Move date utilities to core/platform/date
Move startOfWeek() and quarterIndex() functions from habits domain
usecase package to core/platform/date package, alongside other date
constants and utilities.
